requirements.txt
config
datasets
evaluate
huggingface_hub
ml_collections
nltk
numpy
spacy
torch
tqdm
transformers==4.48.0
wandb
mauve-text
bert-score
rouge
rouge-score
hydra-core==1.3.2
omegaconf==2.3.0
torch-optimi
torch_ema
timm==1.0.9
accelerate==0.29.3